inference-engine/requirements.txt
# Dependencies for the inference-engine project
torch
torchvision
Pillow
pandas
numpy
scikit-learn
opencv-python-headless
matplotlib